+++ apps/router-demo/router-host-vue3-2100/CHANGELOG.md
@@ -1,5 +1,20 @@
 # host-vue3
 
+## 1.0.9
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.8
+- @module-federation/bridge-vue3@0.2.8
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [b00ef13]
+  - @module-federation/enhanced@0.2.7
+  - @module-federation/bridge-vue3@0.2.7
+
 ## 1.0.7
 
 ### Patch Changes